requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.2
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
huggingface-hub==0.12.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
plotly==5.13.0
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1674135640740/work
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch3d==0.7.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.19.3
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1665138452165/work
tenacity==8.2.1
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1672833821273/work
terminado==0.17.1
tifffile==2023.2.3
timm==0.6.11
tinycss2==1.2.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets==5.9.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
zipp==3.14.0
